Reorganize and cleanup BlockChain
